+++ api/routes/index.js
@@ -1,0 +1,5 @@
+const loginRoute = require('./login');
+
+module.exports = function routes(app) {
+  loginRoute(app);
+};
